pyproject.toml
[project]
name = "pydantic-ai-mongodb-query-generator"
version = "1.0.0"
description = "Python UV template with uv, ruff, loguru, and pytest"
authors = [
    {name = "Jag Reehal", email = "jag@jagreehal.com"},
]

requires-python = ">=3.12"
readme = "README.md"
license = {file = "LICENSE"}

dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-ai>=0.0.11",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.5.0",
    "uv>=0.2.0",
    "watchdog>=6.0.0"
]

[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "N",  # flake8-broken-line
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "F401",  # module imported but unused
]

[tool.ruff.lint.isort]
known-first-party = ["python_template"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov=src --cov-report=html -ra -q"
testpaths = [
    "src",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[dependency-groups]
dev = [
    "pytest-asyncio>=0.25.0",
]

